usersのmysqlに関する処理を一箇所に集約
