--- app/src/marvelous/data_store/mysql/__init__.py
@@ -1,1 +1,0 @@
-from .connection import connection
