--- app/src/marvelous/data_store/mysql/users/__init__.py
@@ -1,6 +1,0 @@
-from .get import get_user_by_id, get_users_marvelous_point_ranking
-from .create import create_user
-from .delete import delete_user_by_id
-from .update import (
-    update_user, reset_super_marvelous_left, reset_daily_steps, reset_survival_bonus, reset_marvelous_point
-)
