--- app/src/marvelous/data_store/mysql/users/delete.py
@@ -1,7 +1,0 @@
-from .execute import commit
-
-
-def delete_user_by_id(discord_id: str) -> None:
-    query = "DELETE FROM users WHERE discord_id = %(discord_id)s"
-    params = {"discord_id": discord_id}
-    commit(query, params)
